Decoder input changed to 32 bit unsigned, to make input to function more accessible

+++ boards/lpcxpresso804/driver_examples/gpio/led_output/Layer1.h
@@ -6,6 +6,6 @@ void initPort(void);
 void changeDIR(uint32_t PIN, int direction);
 void initPWM(void);
 void pwmdutycycle(uint32_t percent);
-void decoder(uint8_t value);
+void decoder(uint32_t value);
 
 #endif
